Домашнее задание 1 (исправление) (#2)
Bank

+++ src/TransactionValidator.java
@@ -1,5 +1,5 @@
+import java.math.BigDecimal;
+
 interface TransactionValidator {
-
-    boolean validate(double amount);
-
+    boolean validate(BigDecimal amount);
 }
